fix(payments): amount integer + ép Number khi đọc — chống PostgREST trả bigint dạng chuỗi

+++ payments.sql
@@ -6,7 +6,7 @@ create table if not exists payment_orders (
   email text,
   tier text not null,                      -- starter | pro
   months integer not null default 1,       -- 1 | 12
-  amount bigint not null,                  -- VND
+  amount integer not null,                 -- VND
   status text not null default 'pending',  -- pending | paid | expired
   sepay_tx_id text,
   created_at timestamptz default now(),
@@ -17,7 +17,7 @@ create index if not exists payment_orders_tx_idx on payment_orders (sepay_tx_id)
 
 create table if not exists payment_unmatched (
   id text primary key,                     -- sepay tx id
-  amount bigint,
+  amount integer,
   content text,
   received_at timestamptz default now()
 );
